Move TextEditor to common
