+++ js/src/common/components/TextEditor.js
@@ -1,7 +1,7 @@
-import Component from '../../common/Component';
-import ItemList from '../../common/utils/ItemList';
-import listItems from '../../common/helpers/listItems';
-import Button from '../../common/components/Button';
+import Component from '../Component';
+import ItemList from '../utils/ItemList';
+import listItems from '../helpers/listItems';
+import Button from './Button';
 
 import BasicEditorDriver from '../utils/BasicEditorDriver';
