+++ js/src/forum/components/ComposerBody.js
@@ -1,7 +1,7 @@
 import Component from '../../common/Component';
 import LoadingIndicator from '../../common/components/LoadingIndicator';
 import ConfirmDocumentUnload from '../../common/components/ConfirmDocumentUnload';
-import TextEditor from './TextEditor';
+import TextEditor from '../../common/components/TextEditor';
 import avatar from '../../common/helpers/avatar';
 import listItems from '../../common/helpers/listItems';
 import ItemList from '../../common/utils/ItemList';
