+++ js/src/forum/states/ComposerState.js
@@ -1,7 +1,7 @@
 import subclassOf from '../../common/utils/subclassOf';
 import Stream from '../../common/utils/Stream';
 import ReplyComposer from '../components/ReplyComposer';
-import EditorDriverInterface from '../utils/EditorDriverInterface';
+import EditorDriverInterface from '../../common/utils/EditorDriverInterface';
 
 class ComposerState {
   constructor() {
